New chirpinfer just dropped
